Seperate ssg generation from command
