+++ .gitignore
@@ -352,3 +352,4 @@ MigrationBackup/
 .idea/
 .fleet/
 Samples/*/Vault/
+launchSettings.json
