+++ Nodsoft.MoltenObsidian.Tool/Nodsoft.MoltenObsidian.Tool.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
 		<OutputType>Exe</OutputType>
@@ -15,7 +15,6 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-	  <PackageReference Include="FluentFTP" Version="46.0.2" />
 	  <PackageReference Include="JetBrains.Annotations" Version="2022.3.1" />
 	  <PackageReference Include="Spectre.Console.Cli" Version="0.45.0" />
 	  <PackageReference Include="Throw" Version="1.3.1" />
